+++ src/datatypes/qbsbuildconfigurationkey.ts
@@ -1,7 +1,9 @@
 /** Helper enumeration for wrapping the data keys from the `qbs-configurations.json` file. */
 export enum QbsBuildConfigurationKey {
+    Configurations = 'configurations',
     Description = 'description',
     DisplayName = 'displayName',
     Name = 'name',
     Properties = 'properties',
+    Version = 'version',
 }